requirements.txt
python-amazon-sp-api
SQLAlchemy
psycopg2-binary
tenacity
python-dotenv